+++ src-tauri/src/commands.rs
@@ -6,10 +6,11 @@ use std::{
 use crate::dirstat::DiskItem;
 use crate::models::*;
 use crate::{app::AppState, dirstat::FileInfo};
+
 use serde::{Deserialize, Serialize};
 use tauri::{Manager, State};
 
-use log::{info, trace, warn};
+use log::info;
 use tauri_plugin_shell::ShellExt;
 use ts_rs::TS;
 
@@ -25,57 +26,157 @@ pub fn kill_process(state: State<'_, AppState>, name: String) -> bool {
 }
 
 #[tauri::command]
-pub fn show_folder(path: String) {
+pub fn open(path: String) -> Result<(), String> {
     #[cfg(target_os = "macos")]
     {
-        std::process::Command::new("open")
+        let child = std::process::Command::new("open")
             .arg(path)
             .arg("-R")
-            .spawn()
-            .unwrap();
+            .spawn();
+
+        match child {
+            Ok(_) => Ok(()),
+            Err(e) => Err(e.to_string()),
+        }
     }
     #[cfg(target_os = "linux")]
     {
-        std::process::Command::new("xdg-open")
-            .arg(path)
-            .spawn()
-            .unwrap();
+        let child = std::process::Command::new("xdg-open").arg(path).spawn();
+
+        match child {
+            Ok(_) => Ok(()),
+            Err(e) => Err(e.to_string()),
+        }
     }
     #[cfg(target_os = "windows")]
     {
-        std::process::Command::new("explorer")
-            .arg(path)
-            .spawn()
-            .unwrap();
+        let child = std::process::Command::new("explorer").arg(path).spawn();
+
+        match child {
+            Ok(_) => Ok(()),
+            Err(e) => Err(e.to_string()),
+        }
     }
 }
 
 #[tauri::command]
-pub fn delete_folder(path: String) {
+pub fn show_in_terminal(path: String) -> Result<(), String> {
     #[cfg(target_os = "macos")]
     {
-        std::process::Command::new("rm")
-            .arg("-rf")
+        let child = std::process::Command::new("open")
+            .arg("-a")
+            .arg("Terminal")
             .arg(path)
-            .spawn()
-            .unwrap();
+            .spawn();
+
+        match child {
+            Ok(_) => Ok(()),
+            Err(e) => Err(e.to_string()),
+        }
     }
     #[cfg(target_os = "linux")]
     {
-        std::process::Command::new("rm")
-            .arg("-rf")
-            .arg(path)
-            .spawn()
-            .unwrap();
+        let child = std::process::Command::new("gnome-terminal")
+            .arg("--")
+            .arg("bash")
+            .arg("-c")
+            .arg(format!("cd {}; exec bash", path))
+            .spawn();
+
+        match child {
+            Ok(_) => Ok(()),
+            Err(e) => Err(e.to_string()),
+        }
     }
     #[cfg(target_os = "windows")]
     {
-        std::process::Command::new("del")
-            .arg("/f")
-            .arg("/q")
-            .arg(path)
-            .spawn()
-            .unwrap();
+        let child = std::process::Command::new("cmd")
+            .arg("/c")
+            .arg(format!("start cmd /k cd /d {}", path))
+            .spawn();
+
+        match child {
+            Ok(_) => Ok(()),
+            Err(e) => Err(e.to_string()),
+        }
+    }
+}
+
+#[tauri::command]
+pub fn delete_folder(path: String) -> Result<(), String> {
+    #[cfg(target_os = "macos")]
+    {
+        let child = std::process::Command::new("rm")
+            .arg("-rf")
+            .arg(&path)
+            .spawn();
+
+        match child {
+            Ok(_) => Ok(()),
+            Err(e) => Err(e.to_string()),
+        }
+    }
+    #[cfg(target_os = "linux")]
+    {
+        let child = std::process::Command::new("rm")
+            .arg("-rf")
+            .arg(&path)
+            .spawn();
+
+        match child {
+            Ok(_) => Ok(()),
+            Err(e) => Err(e.to_string()),
+        }
+    }
+    #[cfg(target_os = "windows")]
+    {
+        let child = std::process::Command::new("cmd")
+            .arg("/C")
+            .arg("rmdir")
+            .arg("/S")
+            .arg("/Q")
+            .arg(&path)
+            .spawn();
+
+        match child {
+            Ok(_) => Ok(()),
+            Err(e) => Err(e.to_string()),
+        }
+    }
+}
+
+#[tauri::command]
+pub fn delete_file(path: String) -> Result<(), String> {
+    #[cfg(target_os = "macos")]
+    {
+        let child = std::process::Command::new("rm").arg(&path).spawn();
+
+        match child {
+            Ok(_) => Ok(()),
+            Err(e) => Err(e.to_string()),
+        }
+    }
+    #[cfg(target_os = "linux")]
+    {
+        let child = std::process::Command::new("rm").arg(&path).spawn();
+
+        match child {
+            Ok(_) => Ok(()),
+            Err(e) => Err(e.to_string()),
+        }
+    }
+    #[cfg(target_os = "windows")]
+    {
+        let child = std::process::Command::new("cmd")
+            .arg("/C")
+            .arg("del")
+            .arg(&path)
+            .spawn();
+
+        match child {
+            Ok(_) => Ok(()),
+            Err(e) => Err(e.to_string()),
+        }
     }
 }