+++ src-tauri/src/lib.rs
@@ -22,7 +22,6 @@ mod tray;
 mod utils;
 
 use app::AppState;
-use tauri::menu::Menu;
 
 use std::time::Duration;
 
@@ -71,23 +70,22 @@ fn build_and_run_app(app: AppState) {
 
             Ok(())
         })
-        .on_window_event(|window, event| match event {
-            tauri::WindowEvent::CloseRequested { api, .. } => {
+        .on_window_event(|window, event| {
+            if let tauri::WindowEvent::CloseRequested { api, .. } = event {
                 window.hide().unwrap();
                 api.prevent_close();
             }
-            tauri::WindowEvent::DragDrop { .. } => todo!(),
-
-            _ => {}
         })
         .manage(app)
         .invoke_handler(tauri::generate_handler![
             commands::kill_process,
-            commands::show_folder,
+            commands::open,
+            commands::delete_file,
             commands::delete_folder,
             commands::disk_analysis_flattened,
             commands::add_pachtop_exclusion,
             commands::disk_scan,
+            commands::show_in_terminal,
         ])
         .run(tauri::generate_context!())
         .expect("error while running tauri application");