+++ src/components/area-chart.tsx
@@ -1,10 +1,13 @@
 import HighchartsReact from "highcharts-react-official";
 import * as Highcharts from "highcharts/highstock";
+import HighchartsBoost from "highcharts/modules/boost";
 import { Dispatch, SetStateAction, useEffect, useRef, useState } from "react";
 
 import useSettings from "@/hooks/useSettings";
 import { useMantineTheme } from "@mantine/core";
 import { useViewportSize } from "@mantine/hooks";
+
+HighchartsBoost(Highcharts);
 
 export interface InitialAreaChartStateInput {
   title: {
@@ -27,7 +30,7 @@ export const useAreaChartState = (
   opts: InitialAreaChartStateInput
 ): [Highcharts.Options, Dispatch<SetStateAction<Highcharts.Options>>] => {
   const { other } = useMantineTheme();
-  const { isPerformanceModeEnabled } = useSettings();
+  const { settings } = useSettings();
   const [chartOptions, setChartOptions] = useState<Highcharts.Options>({
     accessibility: {
       enabled: false,
@@ -57,7 +60,7 @@ export const useAreaChartState = (
           enabled: false,
         },
         animation: {
-          duration: isPerformanceModeEnabled ? 0 : 1000,
+          duration: settings.isPerformanceModeEnabled ? 0 : 1000,
         },
       },
 
@@ -180,7 +183,7 @@ export const useAreaChartState = (
       ignoreHiddenSeries: true,
       alignTicks: false,
       backgroundColor: "transparent",
-      animation: isPerformanceModeEnabled ? false : true,
+      animation: settings.isPerformanceModeEnabled ? false : true,
     },
   });