+++ src/components/bar-chart.tsx
@@ -1,11 +1,13 @@
 import * as Highcharts from "highcharts";
 import HighchartsReact from "highcharts-react-official";
+import HighchartsBoost from "highcharts/modules/boost";
 import { Dispatch, SetStateAction, useEffect, useRef, useState } from "react";
 
 import useSettings from "@/hooks/useSettings";
 import { useMantineTheme } from "@mantine/core";
 import { useViewportSize } from "@mantine/hooks";
 
+HighchartsBoost(Highcharts);
 export interface BarChartStatePropsInitialState extends Highcharts.Options {
   custom?: {
     title: string;
@@ -20,13 +22,13 @@ export const useBarChartState = (
   props: BarChartStatePropsInitialState
 ): [BarChartStatePropsInitialState, Dispatch<SetStateAction<BarChartStatePropsInitialState>>] => {
   const { other } = useMantineTheme();
-  const { isPerformanceModeEnabled } = useSettings();
+  const { settings } = useSettings();
   const [chartOptions, setChartOptions] = useState<BarChartStatePropsInitialState>({
     ...props,
 
     chart: {
       type: "column",
-      animation: isPerformanceModeEnabled ? false : true,
+      animation: settings.isPerformanceModeEnabled ? false : true,
       backgroundColor: "transparent",
       borderColor: "transparent",
       style: {
@@ -37,7 +39,7 @@ export const useBarChartState = (
     plotOptions: {
       series: {
         animation: {
-          duration: isPerformanceModeEnabled ? 0 : 1000,
+          duration: settings.isPerformanceModeEnabled ? 0 : 1000,
         },
       },
       column: {