+++ src/features/metrics/components/disks/disk.directory-treeview.tsx
@@ -1,13 +1,16 @@
 import "@/features/metrics/styles/disk-treeview.css";
 
-import { memo, useEffect, useState } from "react";
+import { memo, useCallback, useEffect, useState } from "react";
 import { NodeRendererProps, Tree } from "react-arborist";
 
 import DynamicProgress from "@/components/dynamic-progress";
 import formatBytes from "@/features/metrics/utils/format-bytes";
-import { DiskItem } from "@/lib";
+import { commands, DiskItem } from "@/lib";
+import logger from "@/lib/logger";
+import notification from "@/utils/notification";
 import { Box, Group, Text, useMantineTheme } from "@mantine/core";
 import { IconFile, IconFolderCancel, IconFolderOpen } from "@tabler/icons-react";
+import { Menu, MenuItem } from "@tauri-apps/api/menu";
 
 const iconStyle = { paddingRight: "5px", verticalAlign: "middle" };
 
@@ -24,9 +27,112 @@ const FileIcon = ({ fileName }: { fileName: string }) => {
   }
 };
 
+interface NodeContextMenuProps extends Omit<DiskItem, "children"> {}
+
 const Node = ({ node, style, dragHandle, tree, preview }: NodeRendererProps<DiskItem>) => {
   const { colors } = useMantineTheme();
   const [internalStyles, setInternalStyles] = useState(style);
+
+  const contextMenuOpenAction = useCallback(async (diskItem: NodeContextMenuProps) => {
+    logger.trace("Open action trigger for node: ", diskItem.name);
+
+    commands.open(diskItem.path).catch((err) => {
+      notification.error({
+        title: `Error opening file ${diskItem.name}`,
+        message: "An error occurred while trying to open the file.",
+      });
+      logger.error("Error opening file: ", err);
+    });
+  }, []);
+
+  const contextMenuOpenInExplorerAction = useCallback(async (diskItem: NodeContextMenuProps) => {
+    logger.trace("Open in Explorer action trigger for node: ", diskItem.name);
+
+    commands.open(diskItem.path).catch((err) => {
+      notification.error({
+        title: "Error opening in explorer",
+        message: "An error occurred while trying to open the folder in explorer.",
+      });
+      logger.error("Error opening in explorer: ", err);
+    });
+  }, []);
+
+  const contextMenuShowInTerminalAction = useCallback(async (diskItem: NodeContextMenuProps) => {
+    logger.trace("Show in terminal action trigger for node: ", diskItem.name);
+    commands.showInTerminal(diskItem.path).catch((err) => {
+      notification.error({
+        title: "Error opening in terminal",
+        message: "An error occurred while trying to open the folder in terminal.",
+      });
+      logger.error("Error opening in terminal: ", err);
+    });
+  }, []);
+
+  const contextMenuShowDeleteAction = useCallback(async (diskItem: NodeContextMenuProps) => {
+    logger.trace("Delete action trigger for node: ", diskItem.name);
+
+    if (node.isLeaf) {
+      return commands.deleteFile(diskItem.path).catch((err) => {
+        notification.error({
+          title: `Error deleting file ${diskItem.name}`,
+          message: "An error occurred while trying to delete the file.",
+        });
+        logger.error("Error deleting file: ", err);
+      });
+    }
+
+    commands.deleteFolder(diskItem.path).catch((err) => {
+      notification.error({
+        title: `Error deleting folder ${diskItem.name}`,
+        message: "An error occurred while trying to delete the folder.",
+      });
+      logger.error("Error deleting folder: ", err);
+    });
+  }, []);
+
+  const showContextMenu = useCallback(async (e: React.MouseEvent, diskItem: NodeContextMenuProps) => {
+    logger.trace("Showing context menu for disk item - ", diskItem.name);
+
+    const menuItems = await Promise.all([
+      MenuItem.new({
+        text: "Open",
+        enabled: node.isLeaf,
+        action: () => contextMenuOpenAction(diskItem),
+      }),
+      MenuItem.new({
+        text: "Open in Explorer",
+        enabled: !node.isLeaf, // Disable for files
+        action: () => contextMenuOpenInExplorerAction(diskItem),
+      }),
+
+      MenuItem.new({
+        text: "Open in Terminal",
+        enabled: !node.isLeaf, // Disable for files
+        action: () => contextMenuShowInTerminalAction(diskItem),
+      }),
+      MenuItem.new({
+        text: "Rename (WIP)",
+        action: () => {
+          logger.trace("Rename action trigger for node: ", diskItem.name);
+        },
+      }),
+      MenuItem.new({
+        text: "Copy (WIP)",
+        action: () => {
+          logger.trace("Copy action trigger for node: ", diskItem.name);
+        },
+      }),
+
+      MenuItem.new({
+        text: "Delete",
+        action: () => contextMenuShowDeleteAction(diskItem),
+      }),
+    ]);
+
+    const menu = await Menu.new({ items: menuItems });
+
+    await menu.popup().catch((err) => logger.error("Error showing context menu: ", err));
+  }, []);
 
   useEffect(() => {
     setInternalStyles((prev) => ({
@@ -36,7 +142,17 @@ const Node = ({ node, style, dragHandle, tree, preview }: NodeRendererProps<Disk
   }, [node.isSelected]);
 
   return (
-    <div style={internalStyles} className="node" ref={dragHandle} onClick={() => node.isInternal && node.toggle()}>
+    <div
+      style={internalStyles}
+      className="node"
+      ref={dragHandle}
+      onClick={() => node.isInternal && node.toggle()}
+      onContextMenu={(e) => {
+        const { children, ...prop } = node.data;
+        node.select();
+        showContextMenu(e, prop);
+      }}
+    >
       <Group position="apart" noWrap>
         <div>
           {node.isLeaf ? <FileIcon fileName={node.data.name} /> : <FolderIcon isOpen={node.isOpen} />}
@@ -61,14 +177,53 @@ const Node = ({ node, style, dragHandle, tree, preview }: NodeRendererProps<Disk
 };
 
 interface DiskDirectoryTreeViewProps {
-  data: any;
+  data: DiskItem; // TODO: Fix this type
 }
 const DiskDirectoryTreeView: React.FC<DiskDirectoryTreeViewProps> = (props) => {
   const { data } = props;
 
+  const contextMenuOpenInExplorerAction = useCallback(async (path: string) => {
+    await commands.open(path).catch((err) => {
+      notification.error({
+        title: "Error opening directory",
+        message: "An error occurred while trying to open the directory.",
+      });
+      logger.error("Error opening directory: ", err);
+    });
+  }, []);
+
+  const contextMenuShowInTerminalAction = useCallback(async (path: string) => {
+    commands.showInTerminal(path).catch((err) => {
+      notification.error({
+        title: "Error opening in terminal",
+        message: "An error occurred while trying to open the folder in terminal.",
+      });
+      logger.error("Error opening in terminal: ", err);
+    });
+  }, []);
+  const showContextMenu = useCallback(async (e: React.MouseEvent, path: string) => {
+    const menuItems = await Promise.all([
+      MenuItem.new({
+        text: "Open in Terminal",
+        action: () => contextMenuShowInTerminalAction(path),
+      }),
+      MenuItem.new({
+        text: "Open in Explorer",
+        action: () => contextMenuOpenInExplorerAction(path),
+      }),
+    ]);
+
+    const menu = await Menu.new({ items: menuItems });
+
+    await menu.popup().catch((err) => logger.error("Error showing context menu: ", err));
+  }, []);
   return (
     <Tree
-      data={data}
+      onContextMenu={(e) => {
+        e.preventDefault();
+        showContextMenu(e, data.path);
+      }}
+      data={data.children as DiskItem[]}
       openByDefault={false}
       width={"100%"}
       height={350}